requirements.txt
Flask==2.3.3
flask-cors==3.1.2
openai==1.33.2
youtube-transcript-api==1.2.3
python-dotenv==1.0.0
PyMuPDF==1.22.5
gunicorn==20.1.0